Adds just-in-time Markdown and Yaml parsing.  This should speed up things for when the content and meta functions are not directly called for an entry because they do not have to go through heavy processing.
Also adds a `nocontent` option to pass down through the `Query` class for explicitly telling the `Entry\MarkdownFile` class to not parse content (unless explicitly called).  For this instance, we use `Yaml` directly to grab only the entry frontmatter and convert to metadata.  The `nocontent` option is best used for lists and other front-end output where the content is unnecessary.  It speeds up processing like crazy for long lists of posts and will allow us to do more "widget-style" plugins in the future for various things without bogging a site with lots of 100s or 1,000s of posts down.